+++ spec/rails_app/app/components/js_mixins.rb
@@ -1,0 +1,7 @@
+class JsMixins < Netzke::Base
+  js_configure do |c|
+    c.foo = 100
+    c.mixin :one
+    c.mixin :two
+  end
+end