+++ spec/rails_app/app/components/js_mixins/javascripts/one.js
@@ -1,0 +1,6 @@
+{
+  initComponent: function(){
+    this.callParent();
+    this.title = "One";
+  }
+}