+++ spec/rails_app/app/components/js_mixins/javascripts/two.js
@@ -1,0 +1,6 @@
+{
+  initComponent: function(){
+    this.callParent();
+    this.title = this.title + " Two";
+  }
+}